fix two warnings and a message

+++ shedskin/lib/builtin/tuple.hpp
@@ -181,9 +181,9 @@ template<class T> tuple2<T, T>::tuple2(PyObject *p) {
         throw new TypeError(new str("error in conversion to Shed Skin (tuple expected)"));
 
     this->__class__ = cl_tuple;
-    int size = PyTuple_Size(p);
+    size_t size = PyTuple_Size(p);
     this->units.resize(size);
-    for(int i=0; i<size; i++)
+    for(size_t i=0; i<size; i++)
         this->units.at(i) = __to_ss<T>(PyTuple_GetItem(p, i));
 }
